localstack_task_8/requirements.txt
apache-airflow-providers-amazon
apache-airflow-providers-common-sql
pyspark
pandas
numpy
boto3